Switch to using exact video stream start time
"Exact" may be too strong of a word, as there is a demonstrable delay
between the LQ and HQ versions of the same stream and it is not clear
which version is more correct. Moreover, it is not clear how NicoNico's
Flash player seeks within the video streams. It may seek based on the
time reported in the filename (which is what this commit uses, because
it has the largest precision of all publicly available timestamps,
namely millisecond precision), or it may seek based on the /publish
commands in the getplayerstatus XML (which use the same times truncated
[not rounded] to a whole-second precision), or it may seek based on
something else. Its seeking does not seem very precise either. In fact,
seeking in the Flash player seems to have a random error of up to a few
seconds, which easily trumps all other differences mentioned above.